debug

+++ t5x/examples/scalable_decoder_only/network.py
@@ -216,7 +216,7 @@ class Decoder(nn.Module):
       decoder_mask = layers.make_decoder_mask(
           decoder_target_tokens=decoder_target_tokens,
           #dtype=cfg.dtype,
-          dtype=cfg.dtype,
+          dtype='uint8',
           decoder_causal_attention=decoder_causal_attention,
           decoder_segment_ids=decoder_segment_ids)
